% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acts}
\alias{acts}
\title{THE TRAGEDY OF ROMEO AND JULIET by William Shakespeare
The text as a term document matrix with covariates}
\format{
## `matrix`
A matrix with 788 rows and 5 columns
\describe{
  \item{(Intercept)}{Act 1 is the baseline for the design matrix}
  \item{actstwo}{indicator for Act 2}
  \item{actsthree}{indicator for Act 3}
  \item{actsfour}{indicator for Act 4}
  \item{actsfive}{indicator for Act 5}
  ...
}
}
\source{
<https://www.gutenberg.org/files/1513/1513-h/1513-h.htm>
}
\usage{
acts
}
\description{
THE TRAGEDY OF ROMEO AND JULIET by William Shakespeare
The text as a term document matrix with covariates
}
\keyword{datasets}
