% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acts_continuous}
\alias{acts_continuous}
\title{THE TRAGEDY OF ROMEO AND JULIET by William Shakespeare
The text as a term document matrix with covariates}
\format{
## `vector`
A vector with 788 elements
\describe{
  \item{values}{Continuous conversion of line number as a fraction of the way through each act.}
  ...
}
}
\source{
<https://www.gutenberg.org/files/1513/1513-h/1513-h.htm>
}
\usage{
acts_continuous
}
\description{
THE TRAGEDY OF ROMEO AND JULIET by William Shakespeare
The text as a term document matrix with covariates
}
\keyword{datasets}
