% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_functions.R
\name{beta_distribution_plot}
\alias{beta_distribution_plot}
\title{Internally used plot for the predicted Beta distribution}
\usage{
beta_distribution_plot(betamodel, topic, newdata)
}
\arguments{
\item{betamodel}{the output from betaregression so that betareg::predict
can be used}

\item{topic}{Name of the topic to plot}

\item{newdata}{Design matrix of the new data points at which to predict.}
}
\value{
A plot as produced by ggplot2. Calling this function without
assignment to a variable will send a plot to the plotting window.
}
\description{
Show the predicted Beta distribution at the covariate values.
Intended for internal use.  Used in the function `brett_plot`
}
\examples{
my_input <- create_input(Romeo_and_Juliet_tdm,
                        rownames(Romeo_and_Juliet_tdm),
                        topics = 10,
                        covariates = acts)
my_output <- solve_nmf(my_input)
brett_object <- boot_reg(my_output, 100, model = "OLS",
                 topics = c("death", "love"))
brett_plot(brett_object, topic = "death")

brett_object = get_regression_coefs(my_output, model = "BETA",
                       return_just_coefs = FALSE,
                       topics = c("death", "love"))
brett_plot(brett_object, topic="death") +
              ggplot2::coord_cartesian(ylim = c(0,5))

}
