% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_functions.R
\name{boot_plot}
\alias{boot_plot}
\title{boot_plot}
\usage{
boot_plot(
  boot_samples,
  topic,
  newdata = NULL,
  Model = NULL,
  link = NULL,
  from_stm = FALSE
)
}
\arguments{
\item{boot_samples}{A list of bootstrap samples, as produced by boot_reg().}

\item{topic}{The topic to use as the response variable, labeled by its anchor word.}

\item{newdata}{provided if the bootsamples are regression coefficients and we wish to plot the fitted values at 'newdata' points.  newdata should be a data.frame where the N rows are N new observations with values across the columns.  The rownames are used for plotting as appropriate.}

\item{Model}{Used if 'boot_samples' is not from boot_reg and therefore doesn't contain the model information already.  Used if newdata is supplied so that the function knows how to convert the bootstrap samples into predictions at 'newdata' points.  WILL BE REMOVED IN FUTURE}

\item{link}{Used if 'boot_samples' is not from boot_reg and therefore doesn't contain the model information already.  Used if newdata is supplied so that the function knows how to convert the bootstrap samples into predictions at 'newdata' points.  WILL BE REMOVED IN FUTURE}

\item{from_stm}{logical hack so that the same plots from can be made from the output of STM}
}
\value{
A list as produced by ggplot2. Calling this function without assignment to a variable
will send a plot to the plotting window.
}
\description{
Plot smoothed histograms of regression effects for a given topic.
}
\examples{
my_input <- create_input(Romeo_and_Juliet_tdm,
                        rownames(Romeo_and_Juliet_tdm),
                        topics = 10,
                        covariates = acts)
my_output <- solve_nmf(my_input)
boot_samples <- boot_reg(my_output, 100, Model = "OLS")
boot_plot(boot_samples, topic="death")

}
