% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_functions.R
\name{bootstrap_error_bars}
\alias{bootstrap_error_bars}
\title{Create Error Bars for Bootstrap Inference}
\usage{
bootstrap_error_bars(brett_object, topicvector = NULL, coverage = 0.95)
}
\arguments{
\item{brett_object}{A list of bootstrap samples, as produced by boot_reg().}

\item{topicvector}{the selected topic(s) of interest for inference, default
is all topics, though this is often too much output to be useful.}

\item{coverage}{the level of coverage for the interval, default is 95\%}
}
\value{
A data frame. Each row corresponds to a covariate and the columns
give the CI.
}
\description{
Produce a data frame with two-sided, 95\% confidence intervals
based on bootstrap estimates of sampling distributions.
}
\examples{
my_input = create_input(Romeo_and_Juliet_tdm,
                        rownames(Romeo_and_Juliet_tdm),
                        topics = 10,
                        covariates = acts)
my_output = solve_nmf(my_input)
brett_object = boot_reg(my_output, samples = 100, model = "OLS")
bootstrap_error_bars(brett_object, topicvector = "death")

}
