% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_functions.R
\name{brett_plot}
\alias{brett_plot}
\title{Plots of Output}
\usage{
brett_plot(
  brett_object,
  topic,
  newdata = NULL,
  model = NULL,
  link = NULL,
  from_stm = FALSE
)
}
\arguments{
\item{brett_object}{A list of bootstrap samples, as produced by boot_reg() or
model coefficients from get_regression_coefficients()}

\item{topic}{The topic to use as the response variable,
labeled by its anchor word.}

\item{newdata}{provided if the bootsamples are regression coefficients and we
wish to plot the fitted values at 'newdata' points.  newdata should be a
data.frame where the N rows are N new observations with values across the
columns.  The rownames are used for plotting as appropriate.}

\item{model}{Used if 'boot_samples' is not from boot_reg
and therefore doesn't contain the model information already.  Used if
newdata is supplied so that the  function knows how to convert the
bootstrap samples into predictions at 'newdata' points.
MAY BE REMOVED IN FUTURE}

\item{link}{Used ONLY if 'boot_samples' is not from boot_reg and therefore
doesn't contain the model information already.  Used if newdata is
supplied so that the function knows how to convert the bootstrap
samples into predictions at 'newdata' points.  MAY BE REMOVED IN FUTURE}

\item{from_stm}{logical hack so that the same plots from
can be made from the output of STM}
}
\value{
A list as produced by ggplot2. Calling this function without
assignment to a variable will send a plot to the plotting window.
}
\description{
Plot smoothed histograms of regression effects for a topic.
Use case 1: producing density plots of bootstrap OLS samples with
respect to discrete coefficients.  That's the only time OLS will be valid
anyways.
Use case 2: producing quantile plots for the fitted beta distribution with
respect to covariate coefficients. Works best when the coefficients are
discrete or when "newdata" is provided.
}
\examples{
my_input <- create_input(Romeo_and_Juliet_tdm,
                        rownames(Romeo_and_Juliet_tdm),
                        topics = 10,
                        covariates = acts)
my_output <- solve_nmf(my_input)
brett_object <- boot_reg(my_output, 100, model = "OLS",
                 topics = c("death", "love"))
brett_plot(brett_object, topic = "death")

brett_object = get_regression_coefs(my_output, model = "BETA",
             return_just_coefs = FALSE, topics = c("death", "love"))
brett_plot(brett_object, topic="death") +
              ggplot2::coord_cartesian(ylim = c(0,5))



}
