% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{get_lambda}
\alias{get_lambda}
\title{get_lambda}
\usage{
get_lambda(output)
}
\arguments{
\item{output}{An object of class nmf_output.}
}
\value{
A data frame of the anchor words and the inverse lambda (lambdacrit)
}
\description{
Extracts the inverse lambdas and matches them to the anchor words.
}
\examples{
my_input = create_input(Romeo_and_Juliet_tdm,
                        rownames(Romeo_and_Juliet_tdm),
                        topics = 10,
                        covariates = acts)
my_output = solve_nmf(my_input)
get_lambda(my_output)

}
