% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFregress-package.R
\docType{package}
\name{NMFregress-package}
\alias{NMFregress}
\alias{NMFregress-package}
\title{NMFregress: Beta Regression Effects for Topic Trends}
\description{
Performs Topic Model Decomposition of a text corpus and allows regression effects and hypothesis tests on covariates.
}
\author{
\strong{Maintainer}: Campbell Dave \email{DaveCampbell@cunet.carleton.ca}

Authors:
\itemize{
  \item Phelan Gabe
}

}
\keyword{internal}
