% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{print_top_words}
\alias{print_top_words}
\title{print_top_words}
\usage{
print_top_words(output, n = 10)
}
\arguments{
\item{output}{An object of class nmf_output.}

\item{n}{The number of words to print.}
}
\value{
A list with one element for each topic. These are labeled by anchor
words. Each
list element contains the top words in that topic, printed in order of
importance (largest
pseudo-count in the word-topic matrix.)
}
\description{
Print the most important words associated with each topic.
}
\examples{
my_input = create_input(Romeo_and_Juliet_tdm,
                        rownames(Romeo_and_Juliet_tdm),
                        topics = 10,
                        covariates = acts)
my_output = solve_nmf(my_input)
print_top_words(my_output, n = 25)

}
