% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Romeo_and_Juliet_tdm}
\alias{Romeo_and_Juliet_tdm}
\title{THE TRAGEDY OF ROMEO AND JULIET by William Shakespeare
The text as a term document matrix with covariates}
\format{
## `matrix`
A matrix with 2997 rows and  788 columns:
\describe{
  \item{rows}{unique terms, sorted alphabetically, as defined by the row labels. Terms have been filtered to remove stopwords}
  \item{columns}{Each column is one line of the play}
  \item{elements}{Counts of occurence of each term within each line of the play.}
}
}
\source{
<https://www.gutenberg.org/files/1513/1513-h/1513-h.htm>
}
\usage{
Romeo_and_Juliet_tdm
}
\description{
THE TRAGEDY OF ROMEO AND JULIET by William Shakespeare
The text as a term document matrix with covariates
}
\keyword{datasets}
