% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{solve_nmf}
\alias{solve_nmf}
\title{solve_nmf}
\usage{
solve_nmf(input, user_anchors = NULL)
}
\arguments{
\item{input}{An object of class nmf_input.}

\item{user_anchors}{User-specified anchors. These can be specified if the user wishes to find anchors via their
own method. Defaults to null.}
}
\value{
An object of class nmf_output. This contains the word-topic matrix,
the block of the word-topic matrix associated with anchor words, the anchor words,
the order the anchor words were found, the entire vocabulary (sorted to match the word-topic
matrix), and the number of topics. These are all needed to print a summary of the topics.
}
\description{
Solve the non-negative matrix factorization problem with specified inputs.
}
\examples{
my_input = create_input(Romeo_and_Juliet_tdm,
                        rownames(Romeo_and_Juliet_tdm),
                        topics = 10,
                        covariates = acts)
my_output = solve_nmf(my_input)

}
